requirements.txt
pyvirtualdisplay
selenium
PyYaml
